Decompile func00037f08

+++ src/include/lib/lib_37b00.h
@@ -10,7 +10,7 @@ void func00037e1c(void);
 void func00037e38(void);
 void func00037e68(void);
 u32 func00037ea4(void);
-void func00037f08(u16 arg0, s32 arg1);
+void func00037f08(s32 arg0, s32 arg1);
 void func00037f5c(s32 arg0, s32 arg1);
 u32 func00037fc0(void);
 u32 func00038924(void);
